remove login session globals carry-over
Previously we were capturing some ingo about logged in user in session
globals and carrying it over into the globals in the
init_Sessionglobals.php file.   Eliminated these carry-overs. The
globals and session globals will only be set if it finds an active
faqcebook session.   this will eliminate problems with pictures not
showing correctly due to expired token.  If the facebook token has
expired, we should be logging the user out